+++ src/max_div/_core/_cli/bm_speed/__init__.py
@@ -1,0 +1,3 @@
+from .speed_param import SpeedParam
+
+__all__ = ["SpeedParam"]